requirements.txt
pygame-ce>=2.5.0
pytest>=8.0.0
Pillow>=10.0.0


